systemctl: Avoid possible pointer overflow in loop check
The loop operand `u` is initialized to `unit_infos`, and `c` denotes the
count of unit_infos.  If `unit_infos + c` would overflow, either due to
an excessive `c` or a `unit_infos` near the edge of address space or a
combination thereof, it is possible to reach undefined behavior, because
pointer overflow has undefined behavior according to the C/C++ standards.

Instead, subtract the starting pointer value from the current value of
the loop pointer.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>